Rename VaccineTracker to VaccineAvailability
