+++ src/main/java/org/covid19/vaccinetracker/availability/VaccineAvailability.java
@@ -1,4 +1,4 @@
-package org.covid19.vaccinetracker.tracker;
+package org.covid19.vaccinetracker.availability;
 
 import org.covid19.vaccinetracker.bot.BotService;
 import org.covid19.vaccinetracker.cowin.CowinApiClient;
@@ -10,14 +10,14 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Service
-public class VaccineTracker {
+public class VaccineAvailability {
     private final CowinApiClient cowinApiClient;
     private final VaccinePersistence vaccinePersistence;
     private final VaccineCentersProcessor vaccineCentersProcessor;
     private final BotService botService;
 
-    public VaccineTracker(CowinApiClient cowinApiClient, VaccinePersistence vaccinePersistence,
-                          VaccineCentersProcessor vaccineCentersProcessor, BotService botService) {
+    public VaccineAvailability(CowinApiClient cowinApiClient, VaccinePersistence vaccinePersistence,
+                               VaccineCentersProcessor vaccineCentersProcessor, BotService botService) {
         this.cowinApiClient = cowinApiClient;
         this.vaccinePersistence = vaccinePersistence;
         this.vaccineCentersProcessor = vaccineCentersProcessor;
@@ -39,4 +39,11 @@ public class VaccineTracker {
         log.info("Persisting vaccine availability for pin code {}", pincode);
         vaccinePersistence.persistVaccineCenters(pincode, vaccineCenters);
     }
+
+    public VaccineCenters fetchVaccineAvailabilityFromPersistenceStore(String pincode) {
+        log.info("Fetching Vaccine availability from local store");
+        // botservice.notify(...)
+        return vaccinePersistence.fetchVaccineCentersByPincode(pincode);
+    }
+
 }
