+++ src/main/java/org/covid19/vaccinetracker/cowin/Api.java
@@ -1,9 +1,8 @@
 package org.covid19.vaccinetracker.cowin;
 
+import org.covid19.vaccinetracker.availability.VaccineAvailability;
 import org.covid19.vaccinetracker.model.VaccineCenters;
 import org.covid19.vaccinetracker.notifications.VaccineCentersNotification;
-import org.covid19.vaccinetracker.tracker.VaccineTracker;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,12 +14,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/cowin")
 public class Api {
     private final CowinApiClient cowinApiClient;
-    private final VaccineTracker vaccineTracker;
+    private final VaccineAvailability vaccineAvailability;
     private final VaccineCentersNotification notifications;
 
-    public Api(CowinApiClient cowinApiClient, VaccineTracker vaccineTracker, VaccineCentersNotification notifications) {
+    public Api(CowinApiClient cowinApiClient, VaccineAvailability vaccineAvailability, VaccineCentersNotification notifications) {
         this.cowinApiClient = cowinApiClient;
-        this.vaccineTracker = vaccineTracker;
+        this.vaccineAvailability = vaccineAvailability;
         this.notifications = notifications;
     }
 
@@ -31,12 +30,12 @@ public class Api {
 
     @GetMapping("/fetch/store")
     public ResponseEntity<VaccineCenters> vaccineCentersFromStore(@RequestParam final String pincode) {
-        return ResponseEntity.ok(vaccineTracker.fetchVaccineAvailability(pincode));
+        return ResponseEntity.ok(vaccineAvailability.fetchVaccineAvailabilityFromPersistenceStore(pincode));
     }
 
     @GetMapping("/persist")
     public ResponseEntity<?> tracker(@RequestParam final String pincode) {
-        vaccineTracker.refreshVaccineAvailability(pincode);
+        vaccineAvailability.refreshVaccineAvailability(pincode);
         return ResponseEntity.ok().build();
     }
